Bring windows build to usable state
Some issues remain (ex: file paths)
Linux build untested, may be broken

+++ src/file_io.hpp
@@ -5,7 +5,7 @@
 #include <cstdint>
 #include <cstdlib>
 
-char* loadFile( const char *fname, size_t &fsize );
+char* loadFile( const char *fname, uint64_t &fsize );
 
 uint64_t getFilesize( const char *fname );
